// types.ts
export interface NavItem {
  label: string;
  href: string;
}

export interface ShopInfo {
  id: string;
  name: string;
  address: string;
  tel: string;
  hours: string;
  image: string;
  description: string;
}

export interface NewsItem {
  id: string;
  date: string;
  category: string;
  title: string;
}

export interface ColumnItem {
  id: string;
  date: string;
  category: string;
  title: string;
  image: string;
}